+++ src/lib.rs
@@ -198,6 +198,8 @@ mod config;
 mod create_directory_dialog;
 mod data;
 mod file_dialog;
+/// Information panel showing the preview and metadata of the selected item
+pub mod information_panel;
 mod modals;
 
 pub use config::{