+++ CHANGELOG.md
@@ -6,6 +6,7 @@
 
 - Added `FileDialog::update_with_right_panel_ui` to add a custom right panel to the file dialog [#170](https://github.com/fluxxcode/egui-file-dialog/pull/170) (thanks [@crumblingstatue](https://github.com/crumblingstatue)!)
 - Added `FileDialog::active_selected_entries` and `FileDialog::active_entry` to get information about the current active item/s [#170](https://github.com/fluxxcode/egui-file-dialog/pull/170) (thanks [@crumblingstatue](https://github.com/crumblingstatue)!)
+- Added new `InformationPanel` to optionally display file information and image previews in the right panel (thanks [@hacknus](https://github.com/hacknus) and [@bircni](https://github.com/bircni)!)
 - Added option to show system files in the hamburger menu [#173](https://github.com/fluxxcode/egui-file-dialog/pull/173) (thanks [@crumblingstatue](https://github.com/crumblingstatue)!)
 - Support mapped network devices on Windows [#189](https://github.com/fluxxcode/egui-file-dialog/pull/189)
 - Added the ability to drag and drop files and folders to open their respective path [#192](https://github.com/fluxxcode/egui-file-dialog/pull/192) (thanks [@hacknus](https://github.com/hacknus)!)
@@ -24,13 +25,14 @@
 ### 🔧 Changes
 
 - Use path edit as file to save [#160](https://github.com/fluxxcode/egui-file-dialog/pull/160)
-- Updated sysinfo to version `0.32` [#161](https://github.com/fluxxcode/egui-file-dialog/pull/161)
+- Updated sysinfo to version `0.33` [#220](https://github.com/fluxxcode/egui-file-dialog/pull/220)
 - Made default egui fonts an optional feature `default_fonts` [#163](https://github.com/fluxxcode/egui-file-dialog/pull/163) (thanks [@StarStarJ](https://github.com/StarStarJ)!)
 - Filter directory when loading to improve performance [#169](https://github.com/fluxxcode/egui-file-dialog/pull/169)
 - Implement non blocking directory loading [#177](https://github.com/fluxxcode/egui-file-dialog/pull/177)
 - Only update visible items in the central panel if the search value is empty and the create directory dialog is currently closed [#181](https://github.com/fluxxcode/egui-file-dialog/pull/181)
 - Improve CI [#186](https://github.com/fluxxcode/egui-file-dialog/pull/186) (thanks [@bircni](https://github.com/bircni)!)
 - Use `cmd` for keybindings on MacOS [#205](https://github.com/fluxxcode/egui-file-dialog/pull/205) (thanks [@hacknus](https://github.com/hacknus)!)
+- Cleanup examples [#213](https://github.com/fluxxcode/egui-file-dialog/pull/213) (thanks [@bircni](https://github.com/bircni)!)
 
 ### 📚 Documentation